+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This file follows the best practices from [keepachangelog.com](http://keepachang
 
 ### Added
 
+- Add support for using CDN for dynamic assets (uploaded images, custom compiled stylesheets) when S3 is otherwise in use [#2314](https://github.com/sharetribe/sharetribe/pull/2314)
 - Add possibility to choose between light and dark background image filter for hero and info sections in custom landing pages [#2310](https://github.com/sharetribe/sharetribe/pull/2310)
 
 ## [5.8.0] - 2016-07-15